+++ src/upload.rs
@@ -1,11 +1,9 @@
+use async_graphql::dynamic;
+use async_graphql::dynamic::Type;
+use async_graphql::dynamic::ValueAccessor;
+use async_graphql::Upload;
 use std::borrow::Cow;
 
-use async_graphql::dynamic;
-use async_graphql::dynamic::ValueAccessor;
-use async_graphql::Context;
-use async_graphql::UploadValue;
-
-use crate::errors::InputValueError;
 use crate::errors::InputValueResult;
 use crate::from_value::FromValue;
 use crate::registry::Registry;
@@ -13,49 +11,22 @@ use crate::types::InputTypeName;
 use crate::types::Register;
 use crate::types::TypeName;
 
-pub struct Upload(usize);
-
 impl TypeName for Upload {
     fn get_type_name() -> Cow<'static, str> {
-        "Upload".into()
+        dynamic::TypeRef::UPLOAD.into()
     }
 }
-impl InputTypeName for Upload {}
 
-impl Upload {
-    /// Get the upload value.
-    pub fn value(&self, ctx: &Context<'_>) -> std::io::Result<UploadValue> {
-        ctx.query_env
-            .uploads
-            .get(self.0)
-            .ok_or_else(|| std::io::Error::new(std::io::ErrorKind::Other, "Upload not found"))?
-            .try_clone()
-    }
-}
+impl InputTypeName for Upload {}
 
 impl FromValue for Upload {
     fn from_value(value: async_graphql::Result<ValueAccessor>) -> InputValueResult<Self> {
-        const PREFIX: &str = "#__graphql_file__:";
-        let value = value?;
-        let value = value.string()?;
-
-        if let Some(filename) = value.strip_prefix(PREFIX) {
-            let index = filename.parse::<usize>().map_err(|_| {
-                async_graphql::Error::new(
-                    "Invalid upload value, expected #__graphql_file__:index format",
-                )
-            })?;
-            return Ok(Upload(index));
-        }
-        Err(InputValueError::custom(
-            "Invalid upload value, expected #__graphql_file__:index format",
-        ))
+        Ok(value?.upload()?)
     }
 }
 
 impl Register for Upload {
     fn register(registry: Registry) -> Registry {
-        let upload = dynamic::Scalar::new("Upload");
-        registry.register_type(upload)
+        registry.register_type(Type::Upload)
     }
 }